Add `tanh` sql translation as promised in vignette
Not all the functions in `base_scalar` exist in all DBI connections. For example, none of the hyperbolic trig functions are in postgres (RSQLite gets around this by loading the `extension-functions.c` extension). This commit includes `tanh` since all the other ones are in there and since the documentation says that all of the others are in there.