make explain work with redshift
Redshift EXPLAIN does not accept a (FORMAT ***) statement following it, in contrast to PostgresSQL. This PR allows the user to pass `format = NULL` explicitly to allow `explain()` to work with `redshift`.

I am unable to provide a reproducible example, since I dont have a public instance of a redshift cluster I can connect to.