Soften rules for interpolating values (#1466)
Revert to previous escaping method (i.e. reverts #1368). See https://github.com/bcgov/bcdata/pull/338 for details. If we do this again in the future, I think we have to do argument checking in individual translations. That's a lot of work but allows a whole suite of potentially useful translations for more complex R objects.