Clean up the SQL class (#1691)
This started with removal of `c.sql()` which is confusing because when you mix SQL and non-SQL in a vector, it would escape the non-SQL, using the ambient connection. This lead to confusing behaviour and a very unobvious dependency on the active connection. This fix required a few more explicit calls to `sql()` but was overall straightforward.

I then moved `sql_data_mask()` to labelling variables with `ident()` rather than `sql()` (since they are identifiers!). This revealed the semi joins and case when where we were generating expressions for translation, instead of just the SQL directly.

Then I realised `as.sql()` didn't really do what I expected and instead was used primarily as `as_ident()`. So I removed `as.sql()` which lead me to remove the non-longer used escaping mechanisms for the older ways of specifying tables in schema and catalogs. There were two places that needed to accept either SQL or identifiers, those functions and `win_over()`.