Eliminate `sql_optimise()` (#1720)
`sql_optimise()` was only used for two optimisations (filter + summarise  and arrange + summarise) so it makes sense to eliminate in favour of the optimisation techniques used for the other verbs (i.e. implemented in `add_summarise()`). It turns out that `sql_optimise()` was the wrong level of abstraction because it's hard to recognise potential optimisations once we're dealing with SQL strings.